joitain muutoksia

+++ LabyrintinRatkoja/src/labyrintti/Piirto.java
@@ -15,7 +15,7 @@ public class Piirto {
 
     public Piirto(int koko) {
         this.koko = koko;
-        ruudunKoko = 20 / (koko / 10) + 4;
+        ruudunKoko = (int)(20 / (koko / 10.0) + 4);
         ruudut = RandomizedKruskal.luoLabyrintti(koko);
         ruudut = Muokkaus.kaytavienLisays(ruudut, koko);
     }
